Started moving over contributors
It's tedious stripping off the HTML and also making it look nice. Should
definitely find a way to have the profile pics look nicer, with minimal
CSS, docusaurus doesn't seem to play nice with the existing CSS
(uncomment one of the HTML contributors, and check console errors on the
local site).