Close #93, automatically deploy documentation on push to main
Still to do:
We should decide if we want to put someone's username and email
in deploy.yml.

More info:
Also runs a test deploy when a PR is made to main.
Both actions are based on docusaurus code that's linked in #93.

I tested it on a fork and seems kosher.

That said, I'm not sure how well the actions communicate. For
example, they didn't error when a non-existant directory was
referenced, just created a 404 with no clues.

In addition to the docusurus code, I've added the ability to
manually trigger both the test and the deployment of main, for
whatever good that does.

This might only start working once it's been merged with main.

Closes #93