Add alkiln playground, make subpages and redirect, correct mistakes
Does anyone have time to take a longer look at this? This actually
had to change a bunch of stuff and I'm not sure I caught it all. I
also caught some mistakes I'd made about screenshots of secret variables.
Maybe other things.

Anyway, starting to split this into separate pages, as we've wanted
to do for a while, since there was information I wanted to include that
just didn't belong with the basics.

Closes #337.